+++ CHANGELOG.rst
@@ -11,6 +11,10 @@ Release 2.1.10 (in development)
   division error when creating very small collision shapes (reported
   by Koniption, see issues #3334577 and #3308638).
 
+* Fixed geometry optimizer to handle cases where number of morph
+  vertices does not match number of shape vertices (reported by
+  rlibiez, see issue #3395484).
+
 Release 2.1.9 (26 March 2011)
 =============================